Replace AnagramFindr with more flexible AnagramService
Removed AnagramFindr and corresponding tests as part of a service upgrade. The new AnagramService class offers a more flexible solution for different anagram-solving strategies. Implemented a Command Line Interface (CLI) for users to interact with the service. Furthermore, added a feature to handle specific exceptions related to the operations performed within the service.